fudo_etl/requirements.txt
requests
psycopg2-binary
python-dotenv
# google-cloud-secret-manager # Comentado por ahora para desarrollo local